backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
requests
aiohttp
google-generativeai
tinydb
